projects/31-sanskruti-dhananjay-deshmukh/src/FridgeRAG_Smart_Fridge_Recipe_Assistant-main/fridgerag/requirements.txt
discord.py==2.6.3
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
google-generativeai==0.5.4
groq==0.9.0
aiofiles==23.2.1
apscheduler==3.10.4
httpx==0.27.0
Pillow==10.3.0
python-multipart==0.0.9
